Define sampling rate as long

+++ src/main/java/org/fl/collectionAlbum/format/LossyAlbumAudioFiles.java
@@ -37,7 +37,7 @@ public class LossyAlbumAudioFiles extends AbstractAlbumsAudioFiles {
 
 	private static final String BIT_RATE_TITLE = "Bit rate";
 	
-	public LossyAlbumAudioFiles(ObjectNode audioJson, AudioFileType type, String source, double bitRate, double samplingRate, String note, Set<MediaFilePath> mediaFilePaths) {
+	public LossyAlbumAudioFiles(ObjectNode audioJson, AudioFileType type, String source, double bitRate, long samplingRate, String note, Set<MediaFilePath> mediaFilePaths) {
 		
 		super(audioJson, type, source, samplingRate, note, mediaFilePaths);
 		this.bitRate = bitRate;
